type casting

+++ day5/1.c
@@ -5,6 +5,6 @@ int x,y,n;
 printf("\nENTER X|Y|N ? ");
 scanf("%d%d%d",&x,&y,&n);
 int total;
-total=(x+y+n) / 3;
+total=(float)(x+y+n) / 3;
 printf("\n total = %i \n",total);
 }
